main components fixes
