+++ project/src/components/favourites-empty/favourites-empty.tsx
@@ -1,4 +1,5 @@
 function FavouritesEmpty():JSX.Element {
+
   return (
     <main className="page__main page__main--favorites page__main--favorites-empty">
       <div className="page__favorites-container container">
