+++ project/src/components/main-empty/main-empty.tsx
@@ -1,4 +1,5 @@
 function MainEmpty():JSX.Element {
+
   return (
     <div className="page page--gray page--main">
       <header className="header">
