+++ project/src/components/room/room.tsx
@@ -1,4 +1,5 @@
 function Room():JSX.Element {
+
   return (
     <main className="page__main page__main--property">
       <section className="property">
