+++ project/src/components/sign-in/sign-in.tsx
@@ -1,4 +1,5 @@
 function SignIn(): JSX.Element {
+
   return (
     <div className="page page--gray page--login">
       <header className="header">
